separated StopWatch class.
